+++ testdata-generator-common/src/main/java/io/openepcis/testdata/generator/identifier/instances/GenerateCPI.java
@@ -22,6 +22,7 @@ import io.openepcis.testdata.generator.constants.RandomizationType;
 import io.openepcis.testdata.generator.constants.TestDataGeneratorException;
 import io.openepcis.testdata.generator.format.RandomValueGenerator;
 import io.quarkus.runtime.annotations.RegisterForReflection;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import javax.validation.constraints.NotNull;
@@ -69,11 +70,13 @@ public class GenerateCPI extends GenerateEPC {
           && rangeFrom != null
           && count != null
           && count > 0
-          && rangeFrom >= 0) {
-        for (var rangeID = rangeFrom; rangeID < rangeFrom + count; rangeID++) {
+          && rangeFrom.longValue() >= 0) {
+        for (var rangeID = rangeFrom.longValue();
+            rangeID < rangeFrom.longValue() + count.longValue();
+            rangeID++) {
           formattedCPI.add(CPI_URN_PART + modifiedCPI + "." + rangeID);
         }
-        this.rangeFrom = this.rangeFrom + count;
+        this.rangeFrom = BigInteger.valueOf(this.rangeFrom.longValue() + count.longValue());
       } else if (serialType.equalsIgnoreCase("random") && count != null && count > 0) {
         // Return the list of CPI for RANDOM calculation
         randomMinLength = randomMinLength < 1 || randomMinLength > 12 ? 1 : randomMinLength;
@@ -109,12 +112,14 @@ public class GenerateCPI extends GenerateEPC {
           && rangeFrom != null
           && count != null
           && count > 0
-          && rangeFrom >= 0) {
-        for (var rangeID = rangeFrom; rangeID < rangeFrom + count; rangeID++) {
+          && rangeFrom.longValue() >= 0) {
+        for (var rangeID = rangeFrom.longValue();
+            rangeID < rangeFrom.longValue() + count.longValue();
+            rangeID++) {
           formattedCPI.add(
               DomainName.IDENTIFIER_DOMAIN + CPI_URI_PART + cpi + SERIAL_URI_PART + rangeID);
         }
-        this.rangeFrom = this.rangeFrom + count;
+        this.rangeFrom = BigInteger.valueOf(this.rangeFrom.longValue() + count.longValue());
       } else if (serialType.equalsIgnoreCase("random") && count != null && count > 0) {
         // Return the list of CPI for RANDOM calculation
         randomMinLength = randomMinLength < 1 || randomMinLength > 12 ? 1 : randomMinLength;
